chg size_t vec_mac vec_print + add vec_insert_sort

+++ int_vector_max/int_vector_max.c
@@ -1,10 +1,11 @@
 #include "int_vector_max.h"
+#include <stddef.h>
 
 int int_vector_max(const struct int_vector vec)
 {
     int max = 0;
     
-    for (int i = 0; i < vec.size; i++)
+    for (size_t i = 0; i < vec.size; i++)
     {
         if (data[i] > max)
         {
